+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add configs.core.giantswarm.io CRD.
+
+## [3.17.0] - 2021-02-03
+
+### Changed
+
+- Update CAPZ fork to v0.4.12-alpha1
+
 ## [3.16.1] - 2021-01-28
 
 ### Added
@@ -677,7 +687,8 @@ needing to enter empty strings.
 
 
 
-[Unreleased]: https://github.com/giantswarm/apiextensions/compare/v3.16.1...HEAD
+[Unreleased]: https://github.com/giantswarm/apiextensions/compare/v3.17.0...HEAD
+[3.17.0]: https://github.com/giantswarm/apiextensions/compare/v3.16.1...v3.17.0
 [3.16.1]: https://github.com/giantswarm/apiextensions/compare/v3.16.0...v3.16.1
 [3.16.0]: https://github.com/giantswarm/apiextensions/compare/v3.15.1...v3.16.0
 [3.15.1]: https://github.com/giantswarm/apiextensions/compare/v3.15.0...v3.15.1